backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
sentence-transformers>=2.5.0
faiss-cpu>=1.8.0
numpy>=1.26.0
groq>=0.11.0
langchain-text-splitters>=0.2.0
torch>=2.0.0
